Few updates, may work better.

+++ HardLight/HL_OnLoop.cpp
@@ -7,7 +7,7 @@ void HardLight::OnLoop()
 	Uint32 msCurrent = SDL_GetTicks();
 	if (msCurrent - msPhysics < 1000 / 60) return;
 
-	float elapsed = msCurrent - msPhysics;
+	float elapsed = (float) (msCurrent - msPhysics);
 	elapsed /= 1000.0;
 	gScene->simulate(elapsed);
 	msPhysics = msCurrent;
